- add listings management - check listings ownership

+++ app/Http/Controllers/ListingController.php
@@ -43,6 +43,8 @@ class ListingController extends Controller {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Created listing successfully!');
@@ -54,6 +56,12 @@ class ListingController extends Controller {
     }
 
     public function update(Request $request, Listing $listing) {
+
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -74,8 +82,21 @@ class ListingController extends Controller {
     }
 
     public function destroy(Listing $listing) {
+
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Deleted listing successfully!');
+    }
+
+    // Manage listings
+    public function manage() {
+        return view('listings.manage', [
+            'listings' => auth()->user()->listings
+        ]);
     }
 }
